refactor : supabase 코드 class를 이용해 모듈화하여 간편하게 리팩토링
